+++ mango/im/script.go
@@ -2,6 +2,6 @@ package im
 
 type IM_SCRIPT interface {
 	Init()
-	Update(deltaTime float32)
+	Update(deltaTime float64)
 	Draw()
 }
